% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_attr.R
\name{check_attr}
\alias{check_attr}
\title{Check if object has attribute "map"}
\usage{
check_attr(x)
}
\arguments{
\item{x}{Vector}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Useful with \code{\link[dplyr:mutate_all]{dplyr::mutate_if()}}
}
\examples{
df <- data.frame(cars = 1:3)
attr(df$cars, "map") <-
data.frame(levels = 1:3, labels = c("Volvo", "Saab", "Opel"))
check_attr(df$cars)
}
