% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_vas_exp.R
\docType{data}
\name{eq_vas_exp}
\alias{eq_vas_exp}
\title{Expected EQ VAS dataset}
\format{
A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{res2008_2009}{expected EQ
VAS mean for 2008/2009, numeric} \item{res2010_2011}{expected EQ VAS mean
for 2010/2011, numeric} \item{res2012_2013}{expected EQ VAS mean for
2012/2013, numeric} \item{res2014_2015}{expected EQ VAS mean for 2014/2015,
numeric} }
}
\usage{
eq_vas_exp
}
\description{
A dataset containing expected EQ VAS calculated with a linear model for four
year periods: 2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used
for prom_trends() examples.
}
\keyword{datasets}
