% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_vas_obs.R
\docType{data}
\name{eq_vas_obs}
\alias{eq_vas_obs}
\title{Observed EQ VAS dataset}
\format{
A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{eqvas2008_2009}{observed EQ
VAS mean for 2008/2009, numeric} \item{eqvas2010_2011}{observed EQ VAS mean
for 2010/2011, numeric} \item{eqvas2012_2013}{observed EQ VAS mean for
2012/2013, numeric} \item{eqvas2014_2015}{observed EQ VAS mean for
2014/2015, numeric} }
}
\usage{
eq_vas_obs
}
\description{
A dataset containing observed EQ VAS for four year periods: 2008/2009,
2010/2011, 2012/2013, and 2014/2015. Data set used for prom_trends()
examples.
}
\keyword{datasets}
