% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horisontal_bar_plot.R
\name{guide_axis_label_trans}
\alias{guide_axis_label_trans}
\title{Bar plot function}
\usage{
guide_axis_label_trans(label_trans = identity, ...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=theme_slr]{theme_slr()}}}

\item{df}{Data frame.}

\item{title}{Plot title, \code{NULL} if no title.}

\item{subtitle}{Small text under title, \code{NULL} if no subtitle.}

\item{y_lab}{Y-axis label, use \code{NULL} for no label.}

\item{x_lab}{X-axis label, use \code{NULL} for no label.}
}
\value{
\preformatted{              ggplot object containing bar plot.
}
}
\description{
Plot a bar plot using ggplot2.
}
