% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horisontal_bar_plot.R
\name{guide_train.guide_axis_trans}
\alias{guide_train.guide_axis_trans}
\title{Helper for right-hand side label in figure}
\usage{
\method{guide_train}{guide_axis_trans}(x, ...)
}
\description{
Helper for right-hand side label in figure
}
