% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horisontal_bar_plot.R
\name{horisontal_bar_plot}
\alias{horisontal_bar_plot}
\title{Bar plot function}
\usage{
horisontal_bar_plot(
  df,
  x_var,
  fill_var,
  title = NULL,
  subtitle = NULL,
  y_lab = NULL,
  x_lab = NULL,
  fill_colors = NULL,
  percent_accuracy = 1,
  ...
)
}
\arguments{
\item{df}{Data frame.}

\item{x_var}{Variable for x axis, use string name.
Recommended that \code{x_var} is in character in df.}

\item{fill_var}{Variable for the different colors in bars,
use string name.
Use \code{NULL} if only one color for bars.}

\item{title}{Plot title, \code{NULL} if no title.}

\item{subtitle}{Small text under title, \code{NULL} if no subtitle.}

\item{y_lab}{Y-axis label, use \code{NULL} for no label.}

\item{x_lab}{X-axis label, use \code{NULL} for no label.}

\item{fill_colors}{Color of the different categories in \code{fill_var}.}

\item{percent_accuracy}{Set accuracy for \code{\link[scales:label_percent]{scales::percent_format()}}.}

\item{...}{arguments passed to \code{\link[=theme_slr]{theme_slr()}}}
}
\value{
\preformatted{              ggplot object containing bar plot.
}
}
\description{
Plot a bar plot using ggplot2.
}
