% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot.R
\name{km_plot}
\alias{km_plot}
\title{Kaplan-Meier plot function}
\usage{
km_plot(
  survfit_obj,
  make_step = NULL,
  first_point = 1,
  one_level = FALSE,
  y_lim = NULL,
  percent_accuracy = 1,
  y_breaks = 5,
  x_lim = NULL,
  x_breaks = 1,
  n_points = NULL,
  n_risk_break = 50,
  title = NULL,
  subtitle = NULL,
  y_lab = NULL,
  x_lab = NULL,
  line_size = 0.5,
  show_ci = TRUE,
  ribbon_ci = TRUE,
  ribbon_alpha = 0.5,
  ci_line_size = 0.2,
  line_colors = NULL,
  legend.position = c(0, 1),
  legend_labels = ggplot2::waiver(),
  label_breaks = ggplot2::waiver(),
  ...
)
}
\arguments{
\item{survfit_obj}{Object returned from \code{\link[survival:survfit]{survival::survfit()}}.
Also works with data frame
if year and surv variable exist. lower
and upper variable needed if \code{show_ci = TRUE}.
Specify strata variable needed
if several curves wanted.}

\item{make_step}{If \code{TRUE}, step data will be created.}

\item{first_point}{If \code{make_step = TRUE}, \code{first_point} for KM is
1 and for competing risk 0.}

\item{one_level}{Boolean indicating if there is only one
level in the strata.}

\item{y_lim}{Limit on y-axis.}

\item{percent_accuracy}{Set accuracy for \code{\link[scales:label_percent]{scales::percent_format()}}.}

\item{y_breaks}{Length between each break on y-axis.}

\item{x_lim}{Limit on x-axis.}

\item{x_breaks}{Length between each break on x-axis.}

\item{n_points}{Number of points to be plotted,
useful to change if file is
large because of too many points!}

\item{n_risk_break}{Minimum number at risk to include}

\item{title}{Plot title, \code{NULL} for no title.}

\item{subtitle}{Small text under title, \code{NULL} for no subtitle.}

\item{x_lab, y_lab}{X- and Y-axis labels.}

\item{line_size}{Size of the head lines.}

\item{show_ci}{If \code{TRUE}, show confidence interval lines.}

\item{ribbon_ci}{Show confidence interval}

\item{ribbon_alpha}{Degree of transparency for confidence interval}

\item{ci_line_size}{Size of the confidence interval lines.}

\item{line_colors}{Color of the different curves.}

\item{legend.position}{Position of the legend in plot.}

\item{legend_labels}{Label for each legend key, default order as
they appear in \code{names(survfit_obj$strata)}.}

\item{label_breaks}{Order of the legend keys.}

\item{...}{arguments passed to \code{\link[=theme_slr]{theme_slr()}}}
}
\value{
\preformatted{                   ggplot object containing Kaplan-Meier plot.
}
}
\description{
Kaplan-Meier plot using ggplot2.
}
\examples{
library(survival)

# KM-plot with 2 levels
survfit_obj <- survfit(Surv(time / 365.24, status) ~ sex, survival::colon)
km_plot(survfit_obj, y_lim = c(40, 100), y_breaks = 10, x_lim = c(0, 9))

# KM-plot with 6 levels
survfit_obj <- update(survfit_obj, . ~ . + differ)
km_plot(survfit_obj, y_lim = c(30,100), y_breaks = 10, x_lim = c(0,9),
  line_colors = c('dodgerblue', 'red', 'green', 'black', 'yellow', 'chocolate'))
}
