% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain_exp.R
\docType{data}
\name{pain_exp}
\alias{pain_exp}
\title{Expected pain VAS dataset}
\format{
A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{res2008_2009}{expected pain
VAS mean for 2008/2009, numeric} \item{res2010_2011}{expected pain VAS mean
for 2010/2011, numeric} \item{res2012_2013}{expected pain VAS mean for
2012/2013, numeric} \item{res2014_2015}{expected pain VAS mean for
2014/2015, numeric} }
}
\usage{
pain_exp
}
\description{
A dataset containing expected pain VAS calculated with a linear model for
four year periods: 2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set
used for prom_trends() examples.
}
\keyword{datasets}
