% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain_obs.R
\docType{data}
\name{pain_obs}
\alias{pain_obs}
\title{Observed pain VAS dataset}
\format{
A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{pain2008_2009}{observed
pain VAS mean for 2008/2009, numeric} \item{pain2010_2011}{observed pain
VAS mean for 2010/2011, numeric} \item{pain2012_2013}{observed pain VAS
mean for 2012/2013, numeric} \item{pain2014_2015}{observed pain VAS mean
for 2014/2015, numeric} }
}
\usage{
pain_obs
}
\description{
A dataset containing observed pain VAS for four year periods: 2008/2009,
2010/2011, 2012/2013, and 2014/2015. Data set used for prom_trends()
examples.
}
\keyword{datasets}
