% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_valid.R
\name{pick_valid}
\alias{pick_valid}
\title{Help function for add_prom}
\usage{
pick_valid(x)
}
\arguments{
\item{x}{vector wit hPROM values}
}
\description{
If there are several PROM on same date, this function populate rows with
existing PROM values.
}
