% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satis_obs.R
\docType{data}
\name{satis_obs}
\alias{satis_obs}
\title{Observed satisfaction VAS dataset}
\format{
A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{sati2008_2009}{observed
satisfaction VAS mean for 2008/2009, numeric} \item{sati2010_2011}{observed
satisfaction VAS mean for 2010/2011, numeric} \item{sati2012_2013}{observed
satisfaction VAS mean for 2012/2013, numeric} \item{sati2014_2015}{observed
satisfaction VAS mean for 2014/2015, numeric} }
}
\usage{
satis_obs
}
\description{
A dataset containing observed satisfaction VAS for four year periods:
2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used for
prom_trends() examples.
}
\keyword{datasets}
