% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_order.R
\name{size_order}
\alias{size_order}
\title{Make an ordered factor with most important level first}
\usage{
size_order(x, other_count = NULL, other_level = "Övriga")
}
\arguments{
\item{x}{Vector}

\item{other_count}{How many levels should be considered as other level}

\item{other_level}{What the new level should be called}
}
\description{
Useful when making bar plots and have the most common level in the bottom.
Also possible to remove uncommon levels and create new level called other.
}
\examples{
table(ggplot2::diamonds$cut)
in_size_order <- size_order(ggplot2::diamonds$cut)
table(in_size_order)
in_size_order <- size_order(ggplot2::diamonds$cut, 2, "Others")
table(in_size_order)
}
