% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_colors.R
\name{slr_colors}
\alias{slr_colors}
\title{Colors used by the SLR color pallette}
\usage{
slr_colors(n = NULL)
}
\arguments{
\item{n}{number of colors}
}
\value{
vector with color codes
}
\description{
Colors used by the SLR color pallette
}
\examples{
slr_colors(2) # two colors (not yeallow and blue)
}
