% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slrsave.R
\name{slrsave}
\alias{slrsave}
\title{ggsave function with default settings for SLR annual report}
\usage{
slrsave(
  file = NULL,
  plot = NULL,
  width = 3.14961,
  height = 3.54331,
  units = "in",
  dpi = 600,
  device = "tiff",
  family = "Calibri Light"
)
}
\description{
ggsave function with default settings for SLR annual report
}
