% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_slr.R
\name{theme_slr}
\alias{theme_slr}
\title{The Swedish Arthroplasty Registert theme}
\usage{
theme_slr(
  axis_text_angle = NULL,
  legend.position = "bottom",
  legend.justification = legend.position,
  legend_title = FALSE,
  text_size = 7,
  subtitle_size = 8,
  title_size = 9,
  title_hjust = 0.5,
  subtitle = FALSE,
  x_lab_exists = FALSE,
  title_margin = if (subtitle) 1 else title_size/2
)
}
\arguments{
\item{axis_text_angle}{passed to
\code{axis.text.x = element_text(angle = axis_text_angle)}}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{legend.justification}{anchor point for positioning legend inside plot
("center" or two-element numeric vector) or the justification according to
the plot area when positioned outside the plot}

\item{legend_title}{Include legend title (boolean)}

\item{text_size, title_size, subtitle_size}{Text size for most text, title and subtitle}

\item{title_hjust, title_margin}{passed to
\verb{plot.title = element_text(hjust = title_hjust, margin = margin(b = title_margin)}}

\item{subtitle}{is a subtitle used in the figure (boolean)?}

\item{x_lab_exists}{Indicator if x label should be blank or not}
}
\value{
Modified version of \code{\link[=theme_classic]{theme_classic()}}
}
\description{
Implementation of graphical profile used by the annual report and more.
}
