% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horisontal_bar_plot.R
\name{guide_axis_label_trans}
\alias{guide_axis_label_trans}
\title{Helper functions for work-around with guides on discrete scale.
This solution is presented in issue #3171 at
https://github.com/tidyverse/ggplot2/issues/3171
Credit to user paleolimbot for supplying this solution
Helper for right-hand side label in figure}
\usage{
guide_axis_label_trans(label_trans = identity, ...)
}
\arguments{
\item{label_trans}{labels for axis_guide$label_trans}

\item{...}{arguments passed to guide_axis}
}
\description{
Helper functions for work-around with guides on discrete scale.
This solution is presented in issue #3171 at
https://github.com/tidyverse/ggplot2/issues/3171
Credit to user paleolimbot for supplying this solution
Helper for right-hand side label in figure
}
